[FEAT] 지원자 사진 촬영 구현
