+++ package.json
@@ -32,7 +32,7 @@
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
-  "proxy": "https://192.168.1.104:8080/",
+  "proxy": "https://192.168.0.1:8080/",
   "scripts": {
     "start": "set HTTPS=true&&set SSL_CRT_FILE=./mkcert/cert.pem&&set SSL_KEY_FILE=./mkcert/key.pem&&react-scripts start",
     "build": "react-scripts build",
