+++ src/components/interview/checkSetting/Step3.tsx
@@ -1,159 +1,176 @@
-import React, { useCallback, useEffect, useRef, useState } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import * as S from './Step3.style';
+import { useParams } from 'react-router-dom';
+import axios from 'axios';
+import { PhotoAtom } from '../../../recoil/settingAtomes';
 import { useRecoilState } from 'recoil';
-import { CameraAtom } from '../../../recoil/settingAtomes';
 
 function Step3() {
-    const [recordState, setRecordState] = useState(false); // 녹화 상태(녹화중/중지)
-    const [recordBtn, setRecordBtn] = useState(false); // 녹화 버튼 상태
-    const [recordedMediaUrl, setRecordedMediaUrl] = useState<string | null>(null); // 녹화된 비디오의 URL
-    const videoRef = useRef<HTMLVideoElement>(null);
-    const previewRef = useRef<HTMLVideoElement>(null);
-    const mediaRecorder = useRef<MediaRecorder | null>(null);
-    const videoChunks = useRef<Blob[]>([]);
+  const videoRef = useRef<HTMLVideoElement | null>(null);
+  const canvasRef = useRef<HTMLCanvasElement | null>(null);
 
+  const [canvasState, setCanvasState] = useState('none');
+  const [cameraState, setCameraState] = useState('');
+
+  let { groupId, interviewerId } = useParams<{ groupId: string; interviewerId: string }>();
     // 카메라 테스트 유무 상태
-    const [cameraState, setCameraState] = useRecoilState(CameraAtom);
+    const [photoState, setPhotoState] = useRecoilState(PhotoAtom);
 
-    // 카메라, 마이크 접근 허용
-    const getMediaPermission = useCallback(async () => {
-        try {
-            const constraints = {
-                audio: true,
-                video: true,
-            };
+  useEffect(() => {
+    getWebcam((stream: MediaStream) => {
+      if (videoRef.current) {
+        videoRef.current.srcObject = stream;
+      }
+    });
+  }, []);
 
-            const stream = await navigator.mediaDevices.getUserMedia(constraints);
+  const getWebcam = (callback: (stream: MediaStream) => void) => {
+    try {
+      const constraints = {
+        video: true,
+        audio: false,
+      };
+      navigator.mediaDevices.getUserMedia(constraints).then(callback);
+    } catch (err) {
+      console.log(err);
+    }
+  };
 
-            if (videoRef.current) {
-                videoRef.current.srcObject = stream; // 비디오 요소에 스트림 설정
-                console.log("비디오 스트림 설정됨");
-            } else {
-                console.error("videoRef.current가 null입니다.");
-            }
+  const GoToCamera = () => {
+    setCanvasState('none');
+    setCameraState('');
+    getWebcam((stream: MediaStream) => {
+      if (videoRef.current) {
+        videoRef.current.srcObject = stream;
+      }
+    });
+  };
 
-            const recorder = new MediaRecorder(stream, {
-                mimeType: 'video/webm;codecs=vp9,opus', // 오디오 포함하는 MIME 타입 설정
-            });
+  // 사진 촬영 버튼 클릭 시
+  const sreenShot = () => {
+    setCanvasState('');
+    setCameraState('none');
+    const video = videoRef.current;
+    const canvas = canvasRef.current;
+    if (video && canvas) {
+      const context = canvas.getContext('2d');
+      if (context) {
+        context.scale(-1, 1);
+        context.translate(-canvas.width, 0);
+        context.drawImage(video, 0, 0, canvas.width, canvas.height);
+        canvas.toBlob((blob) => {
+          if (blob) {
+            const file = new File([blob], 'fileName.jpg', { type: 'image/jpeg' });
+            const formData = new FormData();
+            formData.append('file', file);
 
-            recorder.ondataavailable = (e) => {
-                if (e.data.size > 0) {
-                    videoChunks.current.push(e.data); // 데이터가 유효한 경우, Blob 배열에 추가
-                }
-            };
+            // API 호출
+            axios.post(`/interviewGroup/${groupId}/interviewer/${interviewerId}/image`, formData)
+              .then((response) => {
+                //console.log(response.data);
+                setPhotoState(true);
+                // 성공 시 추가 로직 작성
+              })
+              .catch((error) => {
+                console.error('AxiosError:', error);
+                // 실패 시 추가 로직 작성
+              });
+          }
+        }, 'image/jpeg');
 
-            mediaRecorder.current = recorder; // 미디어 레코더 설정
-        } catch (err) {
-            console.error("미디어 장치 접근 중 오류 발생:", err);
-        }
-    }, []);
+        const image = canvas.toDataURL();
+        const link = document.createElement('a');
+        link.href = image;
 
-    useEffect(() => {
-        getMediaPermission(); // 컴포넌트가 마운트될 때 미디어 접근 권한 요청
-    }, [getMediaPermission]);
+        const stream = video.srcObject as MediaStream;
+        stream.getTracks().forEach((track) => {
+          track.stop();
+        });
+      }
+    }
+  };
 
-    // 녹화 중지 시 URL 설정 및 미리보기 준비
-    const handleRecordingStop = async () => {
-        if (mediaRecorder.current) {
-            mediaRecorder.current.stop();
-
-            await new Promise<void>((resolve) => {
-                if (mediaRecorder.current) { // 다시 한 번 null 체크
-                    mediaRecorder.current.onstop = () => {
-                        resolve();
-                    };
-                }
-            });
-
-            if (videoChunks.current.length > 0) {
-                const videoBlob = new Blob(videoChunks.current, { type: 'video/webm' });
-                const videoUrl = URL.createObjectURL(videoBlob);
-                setRecordedMediaUrl(videoUrl);
-
-                console.log("preview: ", previewRef.current);
-                
-                if (previewRef.current) {
-                    previewRef.current.src = videoUrl;
-                    previewRef.current.play();
-                }
-            } else {
-                console.error("비디오 청크가 비어 있습니다.");
-            }
-
-            videoChunks.current = [];
-        }
-    };
-
-    // 녹화 시작/중지 핸들러
-    const handleRecordButtonClick = async () => {
-        if (mediaRecorder.current && recordBtn === false) { // 녹화 시작
-            setRecordBtn(true); // 녹화 버튼 상태를 true로 변경
-            setRecordState(true); // 녹화 상태를 true로 변경
-            mediaRecorder.current.start(); // 녹화 시작
-
-        } else if (mediaRecorder.current && recordBtn === true) { // 녹화 중지
-            setRecordBtn(false); // 녹화 버튼 상태를 false로 변경
-            setRecordState(false); // 녹화 상태를 false로 변경
-            setCameraState(true); // 테스트 유무 true
-            await handleRecordingStop(); // 녹화 중지 후 미리보기 설정
-        }
-    };
-
-    // 다시 녹화 핸들러
-    const handleRetryButtonClick = () => {
-        setCameraState(false); // 테스트 유무 초기화
-        setRecordedMediaUrl(null); // 녹화된 URL 초기화
-        setRecordBtn(false); // 녹화 버튼 상태 초기화
-        setRecordState(false); // 녹화 상태 초기화
-        getMediaPermission(); // 카메라, 마이크 다시 활성화
-    };
-
-    return (
-        <div>
-            <S.StepHeader>카메라, 마이크 점검</S.StepHeader>
-            {!recordedMediaUrl && (
-                <S.Step3Text>
-                    권한 요청 창에서 Google Chrome의 카메라, 마이크 접근 버튼을 클릭한 후, <br />
-                    <S.Step3Txt>"넓은 하늘로의 비상을 꿈꾸며"</S.Step3Txt>를 소리내어 읽어주세요.
-                </S.Step3Text> )}
-            {recordedMediaUrl && (
-                <S.Step3Text>
-                    카메라와 마이크가 잘 연결되었습니다. <br />
-                    재생 버튼을 클릭해 영상이 잘 재생되는지 확인해 주세요. 
-                </S.Step3Text> )}
-            <S.StepMain>
-                {!recordedMediaUrl && (
-                    <S.CameraComponent>
-                        <S.Record><S.RecordTxt recordState={recordState}>●</S.RecordTxt>녹화 중</S.Record>
-                        <S.Camera ref={videoRef} autoPlay border={recordBtn ? '#75A812' : '#D0D2D7'} />
-                        <S.RecordBtnBox>
-                            <S.RecordBtn
-                                btnState={recordBtn}
-                                onClick={handleRecordButtonClick}
-                            >
-                                <S.RecordBtnIcon/>
-                            </S.RecordBtn>
-                        </S.RecordBtnBox>
-                    </S.CameraComponent>
-                )}
-
-                {/* 미리보기 화면 */}
-                {recordedMediaUrl && (
-                    <S.CameraComponent>
-                        <S.Camera ref={previewRef} src={recordedMediaUrl} controls border="#D0D2D7"/>
-                        <S.RecordBtnBox>
-                            <S.ReRecordBtn
-                                btnState={recordBtn}
-                                onClick={handleRetryButtonClick}
-                            >
-                                다시 녹화하기
-                            </S.ReRecordBtn>
-                        </S.RecordBtnBox>
-                    </S.CameraComponent>
-                )}
-            </S.StepMain>
+  return (
+    <div>
+      <S.StepHeader>사진 촬영</S.StepHeader>
+      <S.Step3Text>
+        면접 시 본인 확인을 위한 사진 촬영을 진행해 주세요.<br/>
+        가운데 위치해 촬영 버튼을 클릭해 주세요.
+      </S.Step3Text>
+      <S.CameraDiv>
+        <video 
+          id="videoCam" 
+          ref={videoRef} 
+          autoPlay 
+          style={{ 
+            display: cameraState, 
+            width: "100%", 
+            height: "100%", 
+            objectFit: "cover",
+            borderRadius: "5px"
+          }} 
+        />
+        <canvas 
+          id="canvas" 
+          ref={canvasRef} 
+          width="630" 
+          height="300" 
+          style={{ display: canvasState, width: "630px", height: "300px" }} 
+        />
+      </S.CameraDiv>
+      {canvasState === 'none' ? (
+        <div
+          onClick={sreenShot}
+          style={{
+            display: 'flex',
+            justifyContent: 'center',
+            alignItems: 'center',
+            width: '70px',
+            height: '70px',
+            margin: '10px',
+            borderRadius: '100px',
+            position: 'absolute',
+            zIndex: '101',
+            bottom: '5%',
+            left: '46%',
+            cursor: 'pointer',
+            backgroundColor: 'white',
+          }}
+        >
+          <div
+            style={{
+              textAlign: 'center',
+              width: '60px',
+              height: '60px',
+              border: '2px solid',
+              borderRadius: '100px',
+            }}
+          ></div>
         </div>
-    );
+      ) : (
+        <div 
+          onClick={GoToCamera} 
+          style={{
+            display: "flex", 
+            justifyContent: "center",
+            alignItems: "center",
+            width: "70px",
+            height: "70px",
+            margin: "10px", 
+            borderRadius: "10px",
+            position: "absolute", 
+            zIndex: "101", 
+            bottom: '5%', 
+            left: "46%", 
+            cursor: "pointer", 
+            backgroundColor: "white"
+          }}
+        >
+          <p>다시 촬영</p>
+        </div>
+      )}
+    </div>
+  );
 }
 
 export default Step3;
